Fixed AeronEventLoop; moved tests concerning image liveness to separate test class
