--- reactor-aeron-core/src/test/java/reactor/aeron/TestUtils.java
@@ -1,8 +1,0 @@
-package reactor.aeron;
-
-public class TestUtils {
-
-  public static void log(String msg) {
-    System.out.println(Thread.currentThread().getName() + " - " + msg);
-  }
-}
