+++ reactor-aeron-core/src/test/java/reactor/aeron/UnavailableImageTest.java
@@ -1,0 +1,148 @@
+package reactor.aeron;
+
+import static java.lang.Boolean.TRUE;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import io.aeron.ChannelUriStringBuilder;
+import java.nio.ByteBuffer;
+import java.time.Duration;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
+import java.util.function.Consumer;
+import java.util.function.Function;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.reactivestreams.Publisher;
+import reactor.aeron.AeronOptions.Builder;
+import reactor.aeron.client.AeronClient;
+import reactor.aeron.server.AeronServer;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.ReplayProcessor;
+
+public class UnavailableImageTest extends BaseAeronTest {
+
+  private ChannelUriStringBuilder serverChannel;
+  private ChannelUriStringBuilder clientChannel;
+  private AeronResources aeronResources;
+  private Duration imageLivenessTimeout;
+
+  @BeforeEach
+  void beforeEach() {
+    serverChannel =
+        new ChannelUriStringBuilder()
+            .reliable(TRUE)
+            .media("udp")
+            .endpoint("localhost:" + SocketUtils.findAvailableUdpPort(13000, 14000));
+    clientChannel =
+        new ChannelUriStringBuilder()
+            .reliable(TRUE)
+            .media("udp")
+            .endpoint("localhost:" + SocketUtils.findAvailableUdpPort(14000, 15000));
+    imageLivenessTimeout = Duration.ofSeconds(1);
+    aeronResources =
+        AeronResources.start(
+            AeronResourcesConfig //
+                .builder()
+                .imageLivenessTimeout(imageLivenessTimeout)
+                .build());
+  }
+
+  @AfterEach
+  void afterEach() {
+    if (aeronResources != null) {
+      aeronResources.dispose();
+      aeronResources.onDispose().block(TIMEOUT);
+    }
+  }
+
+  @Test
+  public void testServerDisconnectsSessionAndClientHandleUnavailableImage()
+      throws InterruptedException {
+    ReplayProcessor<ByteBuffer> processor = ReplayProcessor.create();
+    CountDownLatch latch = new CountDownLatch(1);
+
+    createServer(
+        connection -> {
+          connection.onDispose().doOnSuccess(aVoid -> latch.countDown()).subscribe();
+          connection.inbound().receive().subscribe(processor);
+          return connection.onDispose();
+        });
+
+    Connection connection = createConnection();
+    connection
+        .outbound()
+        .send(
+            Flux.range(1, 100)
+                .delayElements(Duration.ofSeconds(1))
+                .map(i -> AeronUtils.stringToByteBuffer("" + i))
+                .log("send"))
+        .then()
+        .subscribe();
+
+    processor.blockFirst();
+
+    connection.dispose();
+
+    latch.await(imageLivenessTimeout.toMillis(), TimeUnit.MILLISECONDS);
+
+    assertEquals(0, latch.getCount());
+  }
+
+  @Test
+  public void testClientClosesSessionAndServerHandleUnavailableImage() throws Exception {
+    OnDisposable server =
+        createServer(
+            connection ->
+                connection
+                    .outbound()
+                    .send(
+                        ByteBufferFlux.from("hello1", "2", "3")
+                            .delayElements(Duration.ofSeconds(1))
+                            .log("server1"))
+                    .then(connection.onDispose()));
+
+    ReplayProcessor<String> processor = ReplayProcessor.create();
+
+    Connection connection =
+        createConnection(
+            options -> {
+              options.clientChannel(clientChannel);
+              options.serverChannel(serverChannel);
+            });
+
+    CountDownLatch latch = new CountDownLatch(1);
+    connection.onDispose().doOnSuccess(aVoid -> latch.countDown()).subscribe();
+
+    connection.inbound().receive().asString().log("client").subscribe(processor);
+
+    processor.take(1).blockLast(Duration.ofSeconds(4));
+
+    server.dispose();
+
+    latch.await(imageLivenessTimeout.toMillis(), TimeUnit.MILLISECONDS);
+
+    assertEquals(0, latch.getCount());
+  }
+
+  private Connection createConnection(Consumer<Builder> options) {
+    return AeronClient.create(aeronResources).options(options).connect().block(TIMEOUT);
+  }
+
+  private OnDisposable createServer(
+      Function<? super Connection, ? extends Publisher<Void>> handler) {
+    return AeronServer.create(aeronResources)
+        .options(options -> options.serverChannel(serverChannel))
+        .handle(handler)
+        .bind()
+        .block(TIMEOUT);
+  }
+
+  private Connection createConnection() {
+    return createConnection(
+        options -> {
+          options.clientChannel(clientChannel);
+          options.serverChannel(serverChannel);
+        });
+  }
+}
